Make Server interactive
If stdin has data, it is read and processed, so we could interpret
some commands here. At the moment, everything aborts the server
which closes #86.

It was necessary to rework the serial code to not wait any longer
for a character received, because a blocking read would prevent
checking stdin.

Tested on Linux and Windows by now, Mac still untested.